Rename xip_ram_perms.cpp and enc_bootloader.cpp to get_....cpp
