+++ get_enc_bootloader.cpp
@@ -5,15 +5,18 @@
 #include <sstream>
 #include <fstream>
 
-#include "xip_ram_perms.h"
-#include "xip_ram_perms_elf.h"
+#include "get_enc_bootloader.h"
+#include "enc_bootloader_elf.h"
+#if HAS_MBEDTLS
+#include "enc_bootloader_mbedtls_elf.h"
+#endif
 
 #include "data_locs.h"
 
 #include "whereami++.h"
 
 
-std::shared_ptr<std::iostream> get_xip_ram_perms() {
+std::shared_ptr<std::iostream> get_enc_bootloader(bool use_mbedtls) {
     // search same directory as executable
     whereami::whereami_path_t executablePath = whereami::getExecutablePath();
     std::string local_loc = executablePath.dirname() + "/";
@@ -22,7 +25,10 @@ std::shared_ptr<std::iostream> get_xip_ram_perms() {
     }
 
     for (auto loc : data_locs) {
-        std::string filename = loc + "xip_ram_perms.elf";
+        std::string filename = loc
+                                + "enc_bootloader"
+                                + (use_mbedtls ? "_mbedtls" : "")
+                                + ".elf";
         std::ifstream i(filename);
         if (i.good()) {
             printf("Picking file %s\n", filename.c_str());
@@ -31,9 +37,16 @@ std::shared_ptr<std::iostream> get_xip_ram_perms() {
         }
     }
 
-    // fall back to embedded xip_ram_perms.elf file
-    printf("Could not find xip_ram_perms.elf file - using embedded binary\n");
+    // fall back to embedded enc_bootloader.elf file
+    printf("Could not find enc_bootloader%s.elf file - using embedded binary\n", use_mbedtls ? "_mbedtls" : "");
     auto tmp = std::make_shared<std::stringstream>();
-    tmp->write(reinterpret_cast<const char*>(xip_ram_perms_elf), xip_ram_perms_elf_SIZE);
+#if HAS_MBEDTLS
+    if (use_mbedtls) {
+        tmp->write(reinterpret_cast<const char*>(enc_bootloader_mbedtls_elf), enc_bootloader_mbedtls_elf_SIZE);
+    } else
+#endif
+    {
+        tmp->write(reinterpret_cast<const char*>(enc_bootloader_elf), enc_bootloader_elf_SIZE);
+    }
     return tmp;
 }
