+++ get_xip_ram_perms.h
